refactor: 이름 나누기 클래스 변경
