+++ src/test/java/ladder/util/NameHelperTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.util;
 
 import org.junit.jupiter.api.Test;
 
@@ -7,14 +7,11 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class UserNamesTest {
-    UserNames userNames;
-
+public class NameHelperTest {
     @Test
     void 쉼표_구분() {
         String names = "buddy,pobi,brown";
-        userNames = new UserNames(names);
         List<String> results = Arrays.asList("buddy", "pobi", "brown");
-        assertThat(userNames.splitNames(names)).isEqualTo(results);
+        assertThat(NameHelper.splitNames(names)).isEqualTo(results);
     }
 }
